requirements.txt
python-dotenv~=1.0.1
asyncpg~=0.29.0